refactor(generator): split module into submodules
